moves bulk actions hook to table grid component
